southcartel/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
cycler==0.10.0
Cython==0.29.24
Django==3.2.5
django-admin-thumbnails==0.2.6
django-jazzmin==2.4.8
django-session-timeout==0.1.0
django-taggit==1.5.1
fuzzywuzzy==0.18.0
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
numpy==1.21.2
pandas==1.3.3
patsy==0.5.2
paypalrestsdk==1.13.1
Pillow==8.3.1
pmdarima==1.8.3
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
python-Levenshtein==0.12.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
statsmodels==0.12.2
threadpoolctl==2.2.0
urllib3==1.26.6
